refactor: change events in classes
