+++ keyboard/src/assets/styles/components/_textarea.scss
@@ -1,4 +1,6 @@
 .textarea {
   width: 100%;
   min-height: 100px;
+
+  outline: none;
 }
