+++ keyboard/src/index.js
@@ -1,5 +1,5 @@
-import {Application} from './assets/js/keyboard'
+import {Application} from './assets/js/application'
 import styles from "./style.scss";
 
 
-const app = new Application()
+const app = new Application(document.body)
