+++ keyboard/src/style.scss
@@ -17,6 +17,10 @@
 //layout
 @import './assets/styles/layouts/container';
 
+.app {
+  height: 100vh;
+}
+
